вынесет утилитарные функции и константы в отдельные модули, добавит экспорты/импорты для модулей
