+++ consts.js
@@ -1,0 +1,6 @@
+export const MAX_KICKS = 6;
+
+export const COUNT_JOLT_MAX = 20;
+export const COUNT_JOLT_MIN = 10;
+export const COUNT_BALL_MAX = 60;
+export const COUNT_BALL_MIN = 20;
