+++ index.html
@@ -69,6 +69,6 @@
 </div>
 <div class="catch_screen"></div>
 
-<script src="main.js"></script>
+<script src="main.js" type="module"></script>
 </body>
 </html>
